Add script for mapping hsv -> ssif
